requirements.txt
langgraph
openai-whisper
spacy
ffmpeg-python
pydub
tqdm
numpy
